refactor: removing excess

+++ src/MyFoodDelivery/Data/Models/OrderDetail.cs
@@ -10,7 +10,7 @@ namespace MyFoodDelivery.Data.Models
         public int Id { get; set; }
         public int OrderId { get; set; }
         public int ProductId { get; set; }
-        public uint Price { get; set; }
+        public double Price { get; set; }
         public virtual Order Order { get; set; }
         public virtual Product Product { get; set; }
     }
